remove 2.2 stuff
