+++ integration-tests/src/test/java/org/openmrs/module/fhir2/providers/r4/MedicationRequestFhirResourceProviderIntegrationTest.java
@@ -9,6 +9,7 @@
  */
 package org.openmrs.module.fhir2.providers.r4;
 
+import static java.nio.charset.StandardCharsets.UTF_8;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.endsWith;
 import static org.hamcrest.Matchers.equalTo;
@@ -18,8 +19,12 @@ import static org.hamcrest.Matchers.instanceOf;
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.notNullValue;
 import static org.hamcrest.Matchers.startsWith;
+import static org.openmrs.module.fhir2.FhirConstants.OPENMRS_FHIR_EXT_MEDICATION_REQUEST_FULFILLER_STATUS;
+import static org.openmrs.module.fhir2.api.util.GeneralUtils.inputStreamToString;
 
+import java.io.InputStream;
 import java.util.List;
+import java.util.Objects;
 
 import lombok.AccessLevel;
 import lombok.Getter;
@@ -39,6 +44,12 @@ public class MedicationRequestFhirResourceProviderIntegrationTest extends BaseFh
 	
 	private static final String WRONG_MEDICATION_REQUEST_UUID = "6d0ae116-0000-4629-9850-f15206e63ab0";
 	
+	private static final String JSON_MERGE_PATCH_MEDICATION_REQUEST_PATH = "org/openmrs/module/fhir2/providers/MedicationRequest_patch.json";
+	
+	private static final String JSON_PATCH_MEDICATION_REQUEST_PATH = "org/openmrs/module/fhir2/providers/MedicationRequest_json_patch.json";
+	
+	private static final String XML_PATCH_MEDICATION_REQUEST_PATH = "org/openmrs/module/fhir2/providers/MedicationRequest_xml_patch.xml";
+	
 	private static final String PATIENT_UUID = "86526ed5-3c11-11de-a0ba-001e3766667a";
 	
 	@Getter(AccessLevel.PUBLIC)
@@ -55,7 +66,7 @@ public class MedicationRequestFhirResourceProviderIntegrationTest extends BaseFh
 	@Test
 	public void shouldReturnExistingMedicationRequestAsJson() throws Exception {
 		MockHttpServletResponse response = get("/MedicationRequest/" + MEDICATION_REQUEST_UUID).accept(FhirMediaTypes.JSON)
-		        .go();
+				.go();
 		
 		assertThat(response, isOk());
 		assertThat(response.getContentType(), is(FhirMediaTypes.JSON.toString()));
@@ -66,6 +77,14 @@ public class MedicationRequestFhirResourceProviderIntegrationTest extends BaseFh
 		assertThat(medicationRequest, notNullValue());
 		assertThat(medicationRequest.getIdElement().getIdPart(), equalTo(MEDICATION_REQUEST_UUID));
 		assertThat(medicationRequest, validResource());
+		
+		// confirm that the new fulfiller extension has been added
+		assertThat(medicationRequest.getExtensionByUrl(OPENMRS_FHIR_EXT_MEDICATION_REQUEST_FULFILLER_STATUS),
+				notNullValue());
+		assertThat(
+				medicationRequest.getExtensionByUrl(OPENMRS_FHIR_EXT_MEDICATION_REQUEST_FULFILLER_STATUS).getValue().toString(),
+				is("RECEIVED"));
+		
 	}
 	
 	@Test
@@ -81,7 +100,7 @@ public class MedicationRequestFhirResourceProviderIntegrationTest extends BaseFh
 	@Test
 	public void shouldReturnExistingMedicationRequestAsXML() throws Exception {
 		MockHttpServletResponse response = get("/MedicationRequest/" + MEDICATION_REQUEST_UUID).accept(FhirMediaTypes.XML)
-		        .go();
+				.go();
 		
 		assertThat(response, isOk());
 		assertThat(response.getContentType(), is(FhirMediaTypes.XML.toString()));
@@ -92,6 +111,13 @@ public class MedicationRequestFhirResourceProviderIntegrationTest extends BaseFh
 		assertThat(medicationRequest, notNullValue());
 		assertThat(medicationRequest.getIdElement().getIdPart(), equalTo(MEDICATION_REQUEST_UUID));
 		assertThat(medicationRequest, validResource());
+		
+		// confirm that the new fulfiller extension has been added
+		assertThat(medicationRequest.getExtensionByUrl(OPENMRS_FHIR_EXT_MEDICATION_REQUEST_FULFILLER_STATUS),
+				notNullValue());
+		assertThat(
+				medicationRequest.getExtensionByUrl(OPENMRS_FHIR_EXT_MEDICATION_REQUEST_FULFILLER_STATUS).getValue().toString(),
+				is("RECEIVED"));
 	}
 	
 	@Test
@@ -102,6 +128,94 @@ public class MedicationRequestFhirResourceProviderIntegrationTest extends BaseFh
 		assertThat(response, isNotFound());
 		assertThat(response.getContentType(), is(FhirMediaTypes.XML.toString()));
 		assertThat(response.getContentAsString(), notNullValue());
+	}
+	
+	@Test
+	public void shouldPatchExistingMedicationRequestUsingJsonMergePatch() throws Exception {
+		String jsonMedicationRequestPatch;
+		try (InputStream is = this.getClass().getClassLoader()
+				.getResourceAsStream(JSON_MERGE_PATCH_MEDICATION_REQUEST_PATH)) {
+			Objects.requireNonNull(is);
+			jsonMedicationRequestPatch = inputStreamToString(is, UTF_8);
+		}
+		
+		MockHttpServletResponse response = patch("/MedicationRequest/" + MEDICATION_REQUEST_UUID)
+				.jsonMergePatch(jsonMedicationRequestPatch).accept(FhirMediaTypes.JSON).go();
+		
+		assertThat(response, isOk());
+		assertThat(response.getContentType(), is(FhirMediaTypes.JSON.toString()));
+		assertThat(response.getContentAsString(), notNullValue());
+		
+		MedicationRequest medicationRequest = readResponse(response);
+		
+		assertThat(medicationRequest, notNullValue());
+		assertThat(medicationRequest.getIdElement().getIdPart(), equalTo(MEDICATION_REQUEST_UUID));
+		assertThat(medicationRequest, validResource());
+		
+		// confirm that the fulfiller extension has been updated
+		assertThat(medicationRequest.getExtensionByUrl(OPENMRS_FHIR_EXT_MEDICATION_REQUEST_FULFILLER_STATUS),
+				notNullValue());
+		assertThat(
+				medicationRequest.getExtensionByUrl(OPENMRS_FHIR_EXT_MEDICATION_REQUEST_FULFILLER_STATUS).getValue().toString(),
+				is("COMPLETED"));
+	}
+	
+	@Test
+	public void shouldPatchExistingMedicationRequestUsingJsonPatch() throws Exception {
+		String jsonMedicationRequestPatch;
+		try (InputStream is = this.getClass().getClassLoader().getResourceAsStream(JSON_PATCH_MEDICATION_REQUEST_PATH)) {
+			Objects.requireNonNull(is);
+			jsonMedicationRequestPatch = inputStreamToString(is, UTF_8);
+		}
+		
+		MockHttpServletResponse response = patch("/MedicationRequest/" + MEDICATION_REQUEST_UUID)
+				.jsonPatch(jsonMedicationRequestPatch).accept(FhirMediaTypes.JSON).go();
+		
+		assertThat(response, isOk());
+		assertThat(response.getContentType(), is(FhirMediaTypes.JSON.toString()));
+		assertThat(response.getContentAsString(), notNullValue());
+		
+		MedicationRequest medicationRequest = readResponse(response);
+		
+		assertThat(medicationRequest, notNullValue());
+		assertThat(medicationRequest.getIdElement().getIdPart(), equalTo(MEDICATION_REQUEST_UUID));
+		assertThat(medicationRequest, validResource());
+		
+		// confirm that the fulfiller extension has been updated
+		assertThat(medicationRequest.getExtensionByUrl(OPENMRS_FHIR_EXT_MEDICATION_REQUEST_FULFILLER_STATUS),
+				notNullValue());
+		assertThat(
+				medicationRequest.getExtensionByUrl(OPENMRS_FHIR_EXT_MEDICATION_REQUEST_FULFILLER_STATUS).getValue().toString(),
+				is("COMPLETED"));
+	}
+	
+	@Test
+	public void shouldPatchExistingMedicationRequestUsingXmlPatch() throws Exception {
+		String xmlMedicationRequestPatch;
+		try (InputStream is = this.getClass().getClassLoader().getResourceAsStream(XML_PATCH_MEDICATION_REQUEST_PATH)) {
+			Objects.requireNonNull(is);
+			xmlMedicationRequestPatch = inputStreamToString(is, UTF_8);
+		}
+		
+		MockHttpServletResponse response = patch("/MedicationRequest/" + MEDICATION_REQUEST_UUID)
+				.xmlPatch(xmlMedicationRequestPatch).accept(FhirMediaTypes.XML).go();
+		
+		assertThat(response, isOk());
+		assertThat(response.getContentType(), is(FhirMediaTypes.XML.toString()));
+		assertThat(response.getContentAsString(), notNullValue());
+		
+		MedicationRequest medicationRequest = readResponse(response);
+		
+		assertThat(medicationRequest, notNullValue());
+		assertThat(medicationRequest.getIdElement().getIdPart(), equalTo(MEDICATION_REQUEST_UUID));
+		assertThat(medicationRequest, validResource());
+		
+		// confirm that the fulfiller extension has been updated
+		assertThat(medicationRequest.getExtensionByUrl(OPENMRS_FHIR_EXT_MEDICATION_REQUEST_FULFILLER_STATUS),
+				notNullValue());
+		assertThat(
+				medicationRequest.getExtensionByUrl(OPENMRS_FHIR_EXT_MEDICATION_REQUEST_FULFILLER_STATUS).getValue().toString(),
+				is("COMPLETED"));
 	}
 	
 	@Test
